requirements.txt
pyrogram
ffmpeg
pytube
youtube-search-python
requests
pillow